refactor(graphic)!: drop GraphicPalette type, port strategy tests to renderer suite
GraphicPalette was a 5-field public struct that the library only consumed
two fields of (center, edge). Background and accent existed solely for
the demo's swatch chip visualization — a demo-driven public API we
shouldn't have shipped.

This commit:
- Deletes Sources/PaletteKit/Graphic/GraphicPalette.swift
- Splits SwatchStrategy out to its own file (Sources/PaletteKit/Graphic/SwatchStrategy.swift)
  since it remains public via Configuration.swatchStrategy
- Deletes Tests/PaletteKitTests/GraphicPaletteTests.swift
- Ports the 6 strategy fallback tests (vibrant/contrast/muted hex checks,
  fallback-when-swatches-missing, nil-swatches, allCases count) to
  PaletteGraphicRendererTests, targeting the internal resolveAnchors
  helper added in the previous commit. background/accent coverage is
  intentionally dropped — those fields are gone from the library.

Demo still references GraphicPalette in CardLabView; the next commit
inlines a private replacement struct. Library + library tests are green:
35/35 (was 43, lost 8 GraphicPalette unit tests, ported 6 to renderer
suite, net −2 since equality / Hashable conformance tests no longer
apply to a non-existent type).

BREAKING CHANGE: `GraphicPalette` public type removed. Callers needing
the resolved (center, edge, background, accent) color set should
replicate the strategy fallback chain locally — see PaletteGraphicRenderer.resolveAnchors
for the canonical center/edge logic. Pre-1.5.0 release window chosen
to absorb breakage while user count is small.
