+++ Tests/PaletteKitTests/PaletteGraphicRendererTests.swift
@@ -86,5 +86,71 @@ struct PaletteGraphicRendererTests {
         )
         #expect(first === second, "expected NSCache to return same CGImage")
     }
+
+    // MARK: - resolveAnchors strategy fallback (ported from GraphicPaletteTests)
+
+    private var fullSwatches: SwatchMap {
+        let swatch = { (c: PaletteColor, role: SwatchRole) in
+            Swatch(color: c, role: role, titleTextColor: .white, bodyTextColor: .white)
+        }
+        return SwatchMap(
+            vibrant:      swatch(PaletteColor(r: 200, g: 80,  b: 40), .vibrant),
+            muted:        swatch(PaletteColor(r: 150, g: 110, b: 90), .muted),
+            darkVibrant:  swatch(PaletteColor(r: 90,  g: 30,  b: 10), .darkVibrant),
+            darkMuted:    swatch(PaletteColor(r: 70,  g: 50,  b: 35), .darkMuted),
+            lightVibrant: swatch(PaletteColor(r: 240, g: 150, b: 110), .lightVibrant),
+            lightMuted:   swatch(PaletteColor(r: 230, g: 200, b: 180), .lightMuted)
+        )
+    }
+
+    @Test("resolveAnchors .vibrant uses vibrant + darkVibrant")
+    func vibrantStrategy() {
+        let (center, edge) = PaletteGraphicRenderer.resolveAnchors(
+            palette: palette, swatches: fullSwatches, strategy: .vibrant
+        )
+        #expect(center.hex == "#c85028")
+        #expect(edge.hex == "#5a1e0a")
+    }
+
+    @Test("resolveAnchors .contrast uses lightVibrant + darkMuted")
+    func contrastStrategy() {
+        let (center, edge) = PaletteGraphicRenderer.resolveAnchors(
+            palette: palette, swatches: fullSwatches, strategy: .contrast
+        )
+        #expect(center.hex == "#f0966e")
+        #expect(edge.hex == "#463223")
+    }
+
+    @Test("resolveAnchors .muted uses muted + darkMuted")
+    func mutedStrategy() {
+        let (center, edge) = PaletteGraphicRenderer.resolveAnchors(
+            palette: palette, swatches: fullSwatches, strategy: .muted
+        )
+        #expect(center.hex == "#966e5a")
+        #expect(edge.hex == "#463223")
+    }
+
+    @Test("resolveAnchors .vibrant falls back to dominant when vibrant swatch is nil")
+    func vibrantFallback() {
+        let (center, edge) = PaletteGraphicRenderer.resolveAnchors(
+            palette: palette, swatches: SwatchMap(), strategy: .vibrant
+        )
+        #expect(center.hex == palette.dominant?.hex)
+        #expect(edge.hex == palette.colors.min(by: { $0.luminance < $1.luminance })?.hex)
+    }
+
+    @Test("resolveAnchors with nil swatches resolves to palette extremes")
+    func nilSwatches() {
+        let (center, edge) = PaletteGraphicRenderer.resolveAnchors(
+            palette: palette, swatches: nil, strategy: .vibrant
+        )
+        #expect(center.hex == palette.dominant?.hex)
+        #expect(edge.hex == palette.colors.min(by: { $0.luminance < $1.luminance })?.hex)
+    }
+
+    @Test("SwatchStrategy.allCases enumerates the three strategies")
+    func allCasesCount() {
+        #expect(SwatchStrategy.allCases.count == 3)
+    }
 }
 #endif
